+++ bfg9000/platforms/core.py
@@ -7,12 +7,15 @@ from pkg_resources import get_entry_info
 from ..objutils import memoize
 from ..versioning import SpecifierSet, Version
 
-__all__ = ['known_platforms', 'parse_triplet', 'Platform', 'PlatformTriplet',
-           'platform_name', 'platform_tuple']
+__all__ = ['known_native_object_formats', 'known_platforms', 'parse_triplet',
+           'Platform', 'PlatformTriplet', 'platform_name', 'platform_tuple']
 
 # This lists the known platform families and genera.
 known_platforms = ['posix', 'windows', 'linux', 'darwin', 'cygwin', 'winnt',
                    'win9x', 'msdos']
+# This lists the known native object formats.
+known_native_object_formats = ['elf', 'mach-o', 'coff']
+
 PlatformTriplet = namedtuple('PlatformTriplet', ['arch', 'vendor', 'sys',
                                                  'abi'])