+++ bfg9000/tools/__init__.py
@@ -1,6 +1,7 @@
 import importlib
 import pkgutil
 
+from ..exceptions import ToolNotFoundError
 from ..objutils import memoize
 
 _builders = {}
@@ -31,7 +32,7 @@ def get_builder(env, lang):
     try:
         fn, multi = _builders[lang]
     except KeyError:
-        raise ValueError('unknown language {!r}'.format(lang))
+        raise ToolNotFoundError('unknown language {!r}'.format(lang))
     return fn(env, lang) if multi else fn(env)
 
 
@@ -48,11 +49,11 @@ def get_tool(env, name):
     try:
         return _tools[name](env)
     except KeyError:
-        raise ValueError('unknown tool {!r}'.format(name))
+        raise ToolNotFoundError('unknown tool {!r}'.format(name))
 
 
 def get_tool_runner(lang):
     try:
         return _tool_runners[lang]
     except KeyError:
-        raise ValueError('unknown tool runner {!r}'.format(lang))
+        raise ToolNotFoundError('unknown tool runner {!r}'.format(lang))