+++ bfg9000/tools/java.py
@@ -10,7 +10,8 @@ with known_langs.make('scala') as x:
     x.vars(compiler='SCALAC', runner='SCALACMD', flags='SCALAFLAGS')
     x.exts(source=['.scala'])
 
-with known_formats.make('jvm', mode='dynamic') as x:
+with known_formats.make('jvm', src_lang='java') as fmt, \
+     fmt.make('dynamic') as x:  # noqa
     x.vars(linker='JAR', flags='JARFLAGS')
 
 _default_cmds = {